REFACTOR: Centralised models list and separated chat input component into 2 files
