+++ types/model.ts
@@ -1,0 +1,9 @@
+/**
+ * Centralised list of AI models available in the application.
+ * Each model has a human-readable label and a provider/model-id value.
+ */
+
+export interface Model {
+  label: string;
+  value: string;
+}
